community/agi-agent-application/cryptocurrency-trading-ai-agent-agishark/code_eng/tools/web2pdf/web2pdf_tool.py
from tools.web2pdf.web2pdf import get_webpage_as_pdf_binary
from agents import function_tool
from typing import Dict, List, Any, Optional
import os
import time
import streamlit as st
from openai import OpenAI

@function_tool
def search_parse_webpage_direct(search_query: str, max_results: int) -> Dict[str, Any]:
    """
    Performs web search, converts search result pages to PDF, and parses documents directly without local storage.
    An integrated tool that processes web search, PDF conversion, and document parsing at once, skipping the local storage step.
    
    Args:
        search_query: Query to search for
        max_results: Maximum number of search results to process
    
    Returns:
        Dict: A dictionary containing the results with the following structure:
            - success (bool): Whether the overall operation was successful
            - results (List): List of parsed document results
            - query (str): The search query used
            - total_processed (int): Number of URLs processed
            - total_success (int): Number of documents successfully parsed
            - error (str): Error message in case of failure
    """
    from tools.document_parser.document_parser import DocumentParser
    
    try:
        print(f"search_parse_webpage_direct called: query='{search_query}', max_results={max_results}")
        
        # Set default value for max_results
        if max_results is None or max_results <= 0:
            max_results = 3
            print(f"Applied default max_results value: 3")
        
        # 1. Perform web search directly using OpenAI API
        # Check API key
        api_key = st.session_state.get('openai_key')
        if not api_key:
            print("OpenAI API key is not set.")
            return {
                'success': False,
                'results': [],
                'query': search_query,
                'total_processed': 0,
                'total_success': 0,
                'error': "OpenAI API key is not set. Please set the key in the API Settings page."
            }
        
        # Perform web search
        urls = []
        try:
            print(f"Starting web search with OpenAI API: query='{search_query}'")
            client = OpenAI(api_key=api_key)
            
            # Execute web search - message-based API call
            response = client.chat.completions.create(
                model="gpt-4o",
                messages=[
                    {"role": "system", "content": "Perform web search and provide accurate information with relevant URLs."},
                    {"role": "user", "content": f"Please search for the following topic: {search_query}"}
                ],
                tools=[
                    {
                        "type": "function",
                        "function": {
                            "name": "web_search",
                            "description": "Performs web search to find information.",
                            "parameters": {
                                "type": "object",
                                "properties": {
                                    "query": {
                                        "type": "string",
                                        "description": "Query to search for"
                                    }
                                },
                                "required": ["query"]
                            }
                        }
                    }
                ]
            )
            
            print(f"Response type: {type(response)}")
            print(f"Response content: {response}")
            
            # Extract URLs from response
            try:
                # Process response format
                if hasattr(response, 'choices') and response.choices:
                    choice = response.choices[0]
                    if hasattr(choice, 'message') and choice.message:
                        message = choice.message
                        print(f"Message content: {message.content if message.content else 'None'}")
                        
                        # 1. Check tool_calls
                        if hasattr(message, 'tool_calls') and message.tool_calls:
                            for tool_call in message.tool_calls:
                                print(f"tool_call: {tool_call}")
                                if hasattr(tool_call, 'function') and tool_call.function:
                                    function = tool_call.function
                                    if hasattr(function, 'arguments') and function.arguments:
                                        try:
                                            import json
                                            args = json.loads(function.arguments)
                                            if 'urls' in args:
                                                urls.extend(args['urls'])
                                        except json.JSONDecodeError:
                                            print(f"JSON parsing failed: {function.arguments}")
                        
                        # 2. Extract URLs from message content
                        if message.content:
                            import re
                            url_pattern = r'https?://[^\s]+'
                            found_urls = re.findall(url_pattern, message.content)
                            if found_urls:
                                print(f"URLs found in content: {found_urls}")
                                urls.extend(found_urls)
                        
                        # 3. Check citation or context
                        if hasattr(message, 'context') and message.context:
                            for ctx_item in message.context:
                                if 'url' in ctx_item:
                                    urls.append(ctx_item['url'])
                
            except Exception as e:
                import traceback
                print(f"Error extracting URLs: {str(e)}\n{traceback.format_exc()}")
            
            # Remove duplicates
            urls = list(dict.fromkeys(urls))
            
            # Limit number of results
            urls = urls[:max_results] if len(urls) > max_results else urls
            print(f"Web search successful: Found {len(urls)} URLs, URLs: {urls}")
            
            # Add hardcoded test URLs (temporary)
            if not urls:
                print("No search results, adding test URLs")
                test_urls = [
                    "https://www.coindesk.com/",
                    "https://cointelegraph.com/",
                    "https://www.bitcoin.com/"
                ]
                urls = test_urls[:max_results]
            
        except Exception as e:
            import traceback
            print(f"Web search failed: {str(e)}\n{traceback.format_exc()}")
            return {
                'success': False,
                'results': [],
                'query': search_query,
                'total_processed': 0,
                'total_success': 0,
                'error': f"Error during web search: {str(e)}"
            }
        
        if not urls:
            print("No search results.")
            return {
                'success': True,
                'results': [],
                'query': search_query,
                'total_processed': 0,
                'total_success': 0,
                'message': "No search results found."
            }
        
        # 2. Convert each URL to PDF binary (without saving)
        pdf_binary_list = []
        processed_count = 0
        
        for url in urls:
            print(f"Converting URL to PDF (not saving): {url}")
            pdf_result = get_webpage_as_pdf_binary(url)
            
            if pdf_result['success']:
                pdf_binary_list.append(pdf_result)
                processed_count += 1
            else:
                print(f"PDF conversion failed: {url}, error: {pdf_result.get('error', 'Unknown error')}")
        
        if not pdf_binary_list:
            return {
                'success': False,
                'results': [],
                'query': search_query,
                'total_processed': len(urls),
                'total_success': 0,
                'error': "Failed to convert all URLs to PDF."
            }
        
        # 3. Pass PDF binary data directly to DocumentParser
        print(f"Starting direct parsing of binary data: {len(pdf_binary_list)} documents")
        parser = DocumentParser()
        parse_result = parser.parse_binary_data(pdf_binary_list)
        
        # 4. Combine results
        parsed_docs = []
        total_success = 0
        
        if parse_result['success']:
            for result in parse_result.get('results', []):
                if result.get('success', False):
                    parsed_docs.append({
                        'success': True,
                        'source_url': result.get('metadata', {}).get('source_url', ''),
                        'text': result.get('text', ''),
                    })
                    total_success += 1
                else:
                    parsed_docs.append({
                        'success': False,
                        'source_url': result.get('source_url', ''),
                        'error': result.get('error', 'Document parsing failed')
                    })
        
        print(f"search_parse_webpage_direct completed: {total_success}/{len(pdf_binary_list)} parsing successful")
        
        return {
            'success': total_success > 0,
            'results': parsed_docs,
            'query': search_query,
            'total_processed': processed_count,
            'total_success': total_success,
            'message': f"Successfully parsed {total_success} out of {processed_count} documents"
        }
        
    except Exception as e:
        import traceback
        error_detail = traceback.format_exc()
        print(f"search_parse_webpage_direct exception occurred: {str(e)}\n{error_detail}")
        
        return {
            'success': False,
            'results': [],
            'query': search_query,
            'total_processed': 0,
            'total_success': 0,
            'error': f"Error during web search and document parsing: {str(e)}"
        } 