community/agi-agent-application/cryptocurrency-trading-ai-agent-agishark/code_eng/requirements.txt
streamlit>=1.44.0
watchdog
anthropic
openai-agents
pandas==2.0.3
numpy==1.24.4
plotly==5.18.0
pyupbit==0.2.33
python-dateutil==2.8.2
anthropic==0.8.1
pytz==2023.3
requests==2.31.0
websocket-client==1.7.0
jinja2==3.1.2
matplotlib==3.7.3

watchdog
webdriver-manager
html5lib
playwright
nest_asyncio
PyMuPDF
Pillow
tweepy
typing
datetime
sys
UPBIT

schedule